Use U64 as the sole integer type in the core.
And temporarily add Bit for quantum stuff.

(Later Bit will be int<1> from the Arithmetic Resource.)

+++ src/builder/tail_loop.rs
@@ -112,7 +112,7 @@ mod test {
         let build_result: Result<Hugr, ValidationError> = {
             let mut loop_b = TailLoopBuilder::new(vec![], vec![BIT], vec![ClassicType::i64()])?;
             let [i1] = loop_b.input_wires_arr();
-            let const_wire = loop_b.add_load_const(Const::i64(1)?)?;
+            let const_wire = loop_b.add_load_const(Const::int(1)?)?;
 
             let break_wire = loop_b.make_break(loop_b.loop_signature()?.clone(), [const_wire])?;
             loop_b.set_outputs(break_wire, [i1])?;
@@ -160,7 +160,7 @@ mod test {
                         let mut branch_1 = conditional_b.case_builder(1)?;
                         let [_b1] = branch_1.input_wires_arr();
 
-                        let wire = branch_1.add_load_const(Const::i64(2)?)?;
+                        let wire = branch_1.add_load_const(Const::int(2)?)?;
                         let break_wire = branch_1.make_break(signature, [wire])?;
                         branch_1.finish_with_outputs([break_wire])?;
